+++ seldom/utils/cache.py
@@ -2,9 +2,15 @@
 seldom cache
 """
 import os
+import sys
 import json
 from seldom.logging import log
 from seldom.utils import file
+
+WINDOWS = True
+if sys.platform != "win32":
+    WINDOWS = False
+    import fcntl
 
 DATA_PATH = os.path.join(file.dir, "cache_data.json")
 
@@ -27,17 +33,30 @@ class Cache:
         :param name: key
         """
         if name is None:
-            with open(DATA_PATH, "w", encoding="utf-8") as json_file:
-                log.info("Clear all cache data")
+            with open(DATA_PATH, "w+", encoding="utf-8") as json_file:
+                if WINDOWS is False:
+                    fcntl.flock(json_file, fcntl.LOCK_EX)
                 json.dump({}, json_file)
+                log.info("Clear all cache data")
+                if WINDOWS is False:
+                    fcntl.flock(json_file, fcntl.LOCK_UN)
         else:
             with open(DATA_PATH, "r+", encoding="utf-8") as json_file:
+                if WINDOWS is False:
+                    fcntl.flock(json_file, fcntl.LOCK_EX)
                 save_data = json.load(json_file)
                 del save_data[name]
+                json.dump(save_data, json_file)
                 log.info(f"Clear cache data: {name}")
+                if WINDOWS is False:
+                    fcntl.flock(json_file, fcntl.LOCK_UN)
 
             with open(DATA_PATH, "w+", encoding="utf-8") as json_file:
+                if WINDOWS is False:
+                    fcntl.flock(json_file, fcntl.LOCK_EX)
                 json.dump(save_data, json_file)
+                if WINDOWS is False:
+                    fcntl.flock(json_file, fcntl.LOCK_UN)
 
     @staticmethod
     def set(data: dict) -> None:
@@ -46,6 +65,8 @@ class Cache:
         :param data:
         """
         with open(DATA_PATH, "r+", encoding="utf-8") as json_file:
+            if WINDOWS is False:
+                fcntl.flock(json_file, fcntl.LOCK_EX)
             save_data = json.load(json_file)
             for key, value in data.items():
                 data = save_data.get(key, None)
@@ -55,8 +76,15 @@ class Cache:
                     log.info(f"update cache data: {key} = {value}")
                 save_data[key] = value
 
+            if WINDOWS is False:
+                fcntl.flock(json_file, fcntl.LOCK_UN)
+
         with open(DATA_PATH, "w+", encoding="utf-8") as json_file:
+            if WINDOWS is False:
+                fcntl.flock(json_file, fcntl.LOCK_EX)
             json.dump(save_data, json_file)
+            if WINDOWS is False:
+                fcntl.flock(json_file, fcntl.LOCK_UN)
 
     @staticmethod
     def get(name=None):
@@ -66,6 +94,8 @@ class Cache:
         :return:
         """
         with open(DATA_PATH, "r+", encoding="utf-8") as json_file:
+            if WINDOWS is False:
+                fcntl.flock(json_file, fcntl.LOCK_EX)
             save_data = json.load(json_file)
             if name is None:
                 return save_data
@@ -73,6 +103,9 @@ class Cache:
             value = save_data.get(name, None)
             if value is not None:
                 log.info(f"Get cache data: {name} = {value}")
+
+            if WINDOWS is False:
+                fcntl.flock(json_file, fcntl.LOCK_UN)
             return value